+++ src/printer.js
@@ -692,6 +692,17 @@ function renderSVG(svg, x, y, pdfKitDoc, fontProvider) {
 	};
 
 	SVGtoPDF(pdfKitDoc, svg.svg, svg.x, svg.y, options);
+
+	if (svg.link) {
+		pdfKitDoc.link(svg.x, svg.y, svg._width, svg._height, svg.link);
+	}
+	if (svg.linkToPage) {
+		pdfKitDoc.ref({Type: 'Action', S: 'GoTo', D: [svg.linkToPage, 0, 0]}).end();
+		pdfKitDoc.annotate(svg.x, svg.y, svg._width, svg._height, { Subtype: 'Link', Dest: [svg.linkToPage - 1, 'XYZ', null, null, null] });
+	}
+	if (svg.linkToDestination) {
+		pdfKitDoc.goTo(svg.x, svg.y, svg._width, svg._height, svg.linkToDestination);
+	}
 }
 
 function beginClip(rect, pdfKitDoc) {